+++ src/Raven.CodeAnalysis/Syntax/InternalSyntax/Parser/Parsers/TypeDeclarationParser.cs
@@ -1256,6 +1256,10 @@ internal class TypeDeclarationParser : SyntaxParser
                     accessibilityKeyword = ReadToken();
                 }
 
+                var scopedKeyword = Token(SyntaxKind.None);
+                if (PeekToken().Kind == SyntaxKind.ScopedKeyword)
+                    scopedKeyword = ReadToken();
+
                 var refKindKeyword = Token(SyntaxKind.None);
                 if (PeekToken().Kind is SyntaxKind.RefKeyword or SyntaxKind.OutKeyword or SyntaxKind.InKeyword)
                     refKindKeyword = ReadToken();
@@ -1309,7 +1313,7 @@ internal class TypeDeclarationParser : SyntaxParser
                     continue;
                 }
 
-                parameterList.Add(Parameter(attributeLists, accessibilityKeyword, refKindKeyword, varParamsKeyword, bindingKeyword, name, null, typeAnnotation, dotDotDotToken, defaultValue));
+                parameterList.Add(Parameter(attributeLists, accessibilityKeyword, scopedKeyword, refKindKeyword, varParamsKeyword, bindingKeyword, name, null, typeAnnotation, dotDotDotToken, defaultValue));
                 parsedParameters++;
             }
 
@@ -1520,6 +1524,10 @@ internal class TypeDeclarationParser : SyntaxParser
                 var parameterStart = Position;
                 var attributeLists = AttributeDeclarationParser.ParseAttributeLists(this);
 
+                var scopedKeyword = Token(SyntaxKind.None);
+                if (PeekToken().Kind == SyntaxKind.ScopedKeyword)
+                    scopedKeyword = ReadToken();
+
                 var refKindKeyword = Token(SyntaxKind.None);
                 if (PeekToken().Kind is SyntaxKind.RefKeyword or SyntaxKind.OutKeyword or SyntaxKind.InKeyword)
                     refKindKeyword = ReadToken();
@@ -1573,7 +1581,7 @@ internal class TypeDeclarationParser : SyntaxParser
                     continue;
                 }
 
-                parameterList.Add(Parameter(attributeLists, Token(SyntaxKind.None), refKindKeyword, varParamsKeyword, bindingKeyword, name, null, typeAnnotation, dotDotDotToken, defaultValue));
+                parameterList.Add(Parameter(attributeLists, Token(SyntaxKind.None), scopedKeyword, refKindKeyword, varParamsKeyword, bindingKeyword, name, null, typeAnnotation, dotDotDotToken, defaultValue));
                 parsedParameters++;
             }